Moved Serialize functions into separate file/Renamed files to service handlers
