+++ cFS/apps/geofence/fsw/src/GeoServiceHandler.c
