+++ cFS/apps/trajectory/fsw/src/TrajServiceHandler.c
